Rebuild cache on legacy files changes
The cached envs were only regenerated on `.tool-versions` files changes,
legacy files like `.nvmrc` and others were properly watched by direnv,
but changes to them were overriden by asdf-direnv caching mechanism.
The current patch adds a check in the cache layer to invalidate the
cache if changes to legacy files are detected. To reduce to the minimum
calls to asdf's slow traversal functions, a last modification timestamp
is added to the cached env file in the form of a comment in the
`watch_file` directives. The check consists then in parsing the known
timestamp from the cached file and compare it to the current
modification time.
Proper hashing (like sha1) could be used instead of modification
timestamps for a more precise caching, but the few false positives
timestamps introduce (the file is saved but its content unchanged) is
enough of a corner case to prefer the slightly faster timestamp
approach.