Rebuild cache on legacy files changes
Cached envs are only regenerated on `.tool-versions` files changes,
legacy files like `.node-version` and others are properly watched by
direnv, but changes to them are overriden by asdf-direnv caching
mechanism.
The current patch adds a list legacy files found in the `.envrc` current
or parent directories + their modification time to the cache file name,
resulting in the cache being properly refreshed when they change.
Instead of relying on asdf to extract a list of files to watch, that
would be pretty slow, an asdf hook is added upon installation that will
generate and keep up-to-date a list of supported legacy files.